refactor: rename Exceptions
